Fix navbar not showing breadcrumbs
Seems like a change in Hugo prevents .GetPage function from finding
pages that had double slash (//) in the begining of the path.

It was a lazy issue which shouldn't have been there from the beginning,
just needed to trim one slash from the start of each path string in nav
fragment.